requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.3.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.12.2
bitarray==2.9.2
black==23.3.0
blinker==1.8.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
ckzg==1.0.2
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coverage==7.2.3
cryptography==42.0.7
cssselect==1.2.0
cycler==0.11.0
cytoolz==0.12.3
Deprecated==1.2.14
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.1.1
filelock==3.11.0
flake8==6.0.0
fonttools==4.39.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
hexbytes==0.3.1
httplib2==0.22.0
identify==2.5.22
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.4
jsonschema==4.17.3
kiwisolver==1.4.4
loguru==0.7.0
lru-dict==1.2.0
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsimonious==0.10.0
pathspec==0.11.1
Pillow==9.3.0
pinecone-client==2.2.1
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
protobuf==4.22.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pydeck==0.9.1
pyflakes==3.0.1
PyGithub==2.3.0
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
pyunormalize==15.1.0
pywin32==306
PyYAML==6.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.7.1
rlp==4.0.1
rsa==4.9
selenium==4.8.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sourcery==1.2.0
streamlit==1.35.0
tenacity==8.3.0
tiktoken==0.3.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typing_extensions==4.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
watchdog==4.0.1
web3==6.19.0
webdriver-manager==3.8.6
websocket-client==1.5.1
websockets==12.0
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
